+++ src/ContaoManager/Plugin.php
@@ -2,9 +2,12 @@
 
 namespace HeimrichHannot\VersionsBundle\ContaoManager;
 
+use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use HeimrichHannot\VersionsBundle\HeimrichHannotVersionsBundle;
+use HeimrichHannot\VersionsBundle\VersionsBundle;
 
 class Plugin implements BundlePluginInterface
 {
@@ -14,8 +17,10 @@ class Plugin implements BundlePluginInterface
     public function getBundles(ParserInterface $parser)
     {
         return [
-            BundleConfig::create('HeimrichHannot\VersionsBundle\VersionsBundle')
-                ->setLoadAfter(['Contao\CoreBundle\ContaoCoreBundle'])
+            BundleConfig::create(VersionsBundle::class)
+                ->setLoadAfter(['Contao\CoreBundle\ContaoCoreBundle']),
+            BundleConfig::create(HeimrichHannotVersionsBundle::class)
+                ->setLoadAfter([ContaoCoreBundle::class]),
         ];
     }
 }