+++ src/Resources/contao/config/config.php
@@ -11,16 +11,20 @@
 /**
  * Add persistent version table names
  */
+
+use HeimrichHannot\VersionsBundle\DataContainer\VersionsContainer;
+use HeimrichHannot\VersionsBundle\Model\VersionModel;
+
 array_insert($GLOBALS['PERSISTENT_VERSION_TABLES'], 0, []);
 
 /**
  * Purge jobs
  */
 $GLOBALS['TL_PURGE']['tables']['versions'] = [
-    'callback' => ['HeimrichHannot\Versions\Automator', 'purgeVersionTable'],
+    'callback' => [VersionsContainer::class, 'purgeTable'],
     'affected' => ['tl_version']];
 
 /**
  * Models
  */
-$GLOBALS['TL_MODELS']['tl_version'] = 'HeimrichHannot\Versions\VersionModel';
+$GLOBALS['TL_MODELS']['tl_version'] = VersionModel::class;