+++ src/VersionsBundle.php
@@ -4,6 +4,12 @@ namespace HeimrichHannot\VersionsBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
+/**
+ * Class VersionsBundle
+ * @package HeimrichHannot\VersionsBundle
+ *
+ * @deprecated Use HeimrichHannotVersionsBundle class instead
+ */
 class VersionsBundle extends Bundle
 {
 }